Moved fecthing logic to hooks
