+++ src/components/AuthLoading.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 
 import { useUser } from '../shared/hooks/useUser';
-import { Redirect } from 'react-router';
+import { Redirect } from 'react-router-dom';
 
 interface IAuthLoadingProps {
   to?: string;
 }
 
 export const AuthLoading: (props: IAuthLoadingProps) => JSX.Element = ({ to }) => {
-  const [user, hasLoaded] = useUser();
+  const [user] = useUser();
   const { id } = user;
 
   return (
